+++ packages/slate-react/package.json
@@ -25,9 +25,11 @@
   },
   "devDependencies": {
     "@babel/runtime": "^7.7.4",
+    "@types/jest": "^27.4.1",
+    "@types/jsdom": "^16.2.14",
     "@types/react": "^16.9.13",
     "@types/react-dom": "^16.9.4",
-    "jsdom": "^16.6.0",
+    "@types/react-test-renderer": "^17.0.1",
     "react": ">=16.8.0",
     "react-dom": ">=16.8.0",
     "react-test-renderer": ">=16.8.0",
