+++ packages/slate-react/test/tsconfig.json
@@ -1,0 +1,4 @@
+{
+  "extends": "../../../config/typescript/tsconfig.json",
+  "references": [{ "path": "../" }]
+}
